Beta v2.1
Main Menu working, Level 1 fixed to include a quit button and a better respawn button.
A template was created to create levels off of for quicker level creating in the future.
A "pause" popup was added so that when you start or respawn in a level you have to press Space to start moving.